Hazard Items Swapped around
hazard items swapped around
1 -> 1
2 -> 3
3 -> 4
4 -> 2

blizzard bigger
cracked brick wall wider
