+++ GGK/Assets/Prefabs/ItemPrefabs/Hazard/HazardTier2.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &8921437432877574983
+--- !u!1 &5949921728926908992
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,13 +8,13 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 7170534580459787604}
-  - component: {fileID: 3799705688799236530}
-  - component: {fileID: 4775784872633617760}
-  - component: {fileID: 4032552068930777776}
-  - component: {fileID: 3722686693565726986}
-  - component: {fileID: -6945963496115081675}
-  - component: {fileID: -2238994353808586959}
+  - component: {fileID: 6181112050242537461}
+  - component: {fileID: 7437612702298188028}
+  - component: {fileID: 5872779163969093666}
+  - component: {fileID: 1330375094248605411}
+  - component: {fileID: 3245099375910505891}
+  - component: {fileID: 3641778909247814450}
+  - component: {fileID: 7721465102789523379}
   m_Layer: 3
   m_Name: HazardTier2
   m_TagString: Hazard
@@ -22,36 +22,36 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &7170534580459787604
+--- !u!4 &6181112050242537461
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8921437432877574983}
+  m_GameObject: {fileID: 5949921728926908992}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 134.42183, y: -32.000008, z: -13.634865}
-  m_LocalScale: {x: 6, y: 6, z: 1}
-  m_ConstrainProportionsScale: 0
+  m_LocalPosition: {x: 126.21524, y: -33.750008, z: -11.912199}
+  m_LocalScale: {x: 2.5, y: 2.5, z: 2.5}
+  m_ConstrainProportionsScale: 1
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!33 &3799705688799236530
+--- !u!33 &7437612702298188028
 MeshFilter:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8921437432877574983}
+  m_GameObject: {fileID: 5949921728926908992}
   m_Mesh: {fileID: 10202, guid: 0000000000000000e000000000000000, type: 0}
---- !u!23 &4775784872633617760
+--- !u!23 &5872779163969093666
 MeshRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8921437432877574983}
+  m_GameObject: {fileID: 5949921728926908992}
   m_Enabled: 1
   m_CastShadows: 1
   m_ReceiveShadows: 1
@@ -65,7 +65,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 2100000, guid: b9adce2efd3460c4db3379d04d30be3f, type: 2}
+  - {fileID: 2100000, guid: 112bb6e938500964587c6f7d6bff95c7, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
@@ -87,13 +87,13 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!65 &4032552068930777776
+--- !u!65 &1330375094248605411
 BoxCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8921437432877574983}
+  m_GameObject: {fileID: 5949921728926908992}
   m_Material: {fileID: 0}
   m_IncludeLayers:
     serializedVersion: 2
@@ -103,18 +103,18 @@ BoxCollider:
     m_Bits: 0
   m_LayerOverridePriority: 0
   m_IsTrigger: 1
-  m_ProvidesContacts: 1
+  m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 3
   m_Size: {x: 1, y: 1, z: 1}
   m_Center: {x: 0, y: 0, z: 0}
---- !u!54 &3722686693565726986
+--- !u!54 &3245099375910505891
 Rigidbody:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8921437432877574983}
+  m_GameObject: {fileID: 5949921728926908992}
   serializedVersion: 4
   m_Mass: 1
   m_Drag: 0
@@ -135,19 +135,19 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 126
   m_CollisionDetection: 0
---- !u!114 &-6945963496115081675
+--- !u!114 &3641778909247814450
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8921437432877574983}
+  m_GameObject: {fileID: 5949921728926908992}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: d5a57f767e5e46a458fc5d3c628d0cbb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  GlobalObjectIdHash: 174807141
+  GlobalObjectIdHash: 609584689
   InScenePlacedSourceGlobalObjectIdHash: 0
   AlwaysReplicateAsRoot: 0
   SynchronizeTransform: 1
@@ -156,20 +156,20 @@ MonoBehaviour:
   SpawnWithObservers: 1
   DontDestroyWithOwner: 0
   AutoObjectParentSync: 1
---- !u!114 &-2238994353808586959
+--- !u!114 &7721465102789523379
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8921437432877574983}
+  m_GameObject: {fileID: 5949921728926908992}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d77190dbf4a2255408a880a3bd8dd19e, type: 3}
+  m_Script: {fileID: 11500000, guid: 8044fcaf0c56b1f448367db2d69bfa64, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   timer: 0
-  rb: {fileID: 3722686693565726986}
+  rb: {fileID: 3245099375910505891}
   itemCategory: Hazard
   useCount: 1
   isTimed: 0
