+++ GGK/Assets/Prefabs/ItemPrefabs/Hazard/HazardTier2.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a14434a4591835e49b0ac4efed41f8d4
+guid: 55aa4467dffa2c24e83ee339dc6ba070
 PrefabImporter:
   externalObjects: {}
   userData:
