+++ GGK/Assets/Prefabs/ItemPrefabs/Hazard/HazardTier3.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &8053848096856963621
+--- !u!1 &8921437432877574983
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,13 +8,13 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 2161231715911477302}
-  - component: {fileID: 5122202980572606189}
-  - component: {fileID: 580488047461183586}
-  - component: {fileID: 6013441621889927310}
-  - component: {fileID: 2075625805928030967}
-  - component: {fileID: -7847993167158863882}
-  - component: {fileID: -2925161246315457296}
+  - component: {fileID: 7170534580459787604}
+  - component: {fileID: 3799705688799236530}
+  - component: {fileID: 4775784872633617760}
+  - component: {fileID: 4032552068930777776}
+  - component: {fileID: 3722686693565726986}
+  - component: {fileID: -6945963496115081675}
+  - component: {fileID: -2238994353808586959}
   m_Layer: 3
   m_Name: HazardTier3
   m_TagString: Hazard
@@ -22,36 +22,36 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &2161231715911477302
+--- !u!4 &7170534580459787604
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8053848096856963621}
+  m_GameObject: {fileID: 8921437432877574983}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 138.34431, y: -32.50001, z: -17.89946}
-  m_LocalScale: {x: 2.5, y: 2.5, z: 2.5}
-  m_ConstrainProportionsScale: 1
+  m_LocalPosition: {x: 0, y: 3, z: 0}
+  m_LocalScale: {x: 12, y: 6, z: 1}
+  m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!33 &5122202980572606189
+--- !u!33 &3799705688799236530
 MeshFilter:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8053848096856963621}
-  m_Mesh: {fileID: 10206, guid: 0000000000000000e000000000000000, type: 0}
---- !u!23 &580488047461183586
+  m_GameObject: {fileID: 8921437432877574983}
+  m_Mesh: {fileID: 10202, guid: 0000000000000000e000000000000000, type: 0}
+--- !u!23 &4775784872633617760
 MeshRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8053848096856963621}
+  m_GameObject: {fileID: 8921437432877574983}
   m_Enabled: 1
   m_CastShadows: 1
   m_ReceiveShadows: 1
@@ -65,7 +65,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 2100000, guid: cda38090a9c1cb74f976f018e305f2fa, type: 2}
+  - {fileID: 2100000, guid: b9adce2efd3460c4db3379d04d30be3f, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
@@ -87,13 +87,13 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!65 &6013441621889927310
+--- !u!65 &4032552068930777776
 BoxCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8053848096856963621}
+  m_GameObject: {fileID: 8921437432877574983}
   m_Material: {fileID: 0}
   m_IncludeLayers:
     serializedVersion: 2
@@ -103,18 +103,18 @@ BoxCollider:
     m_Bits: 0
   m_LayerOverridePriority: 0
   m_IsTrigger: 1
-  m_ProvidesContacts: 0
+  m_ProvidesContacts: 1
   m_Enabled: 1
   serializedVersion: 3
   m_Size: {x: 1, y: 1, z: 1}
   m_Center: {x: 0, y: 0, z: 0}
---- !u!54 &2075625805928030967
+--- !u!54 &3722686693565726986
 Rigidbody:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8053848096856963621}
+  m_GameObject: {fileID: 8921437432877574983}
   serializedVersion: 4
   m_Mass: 1
   m_Drag: 0
@@ -135,19 +135,19 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 126
   m_CollisionDetection: 0
---- !u!114 &-7847993167158863882
+--- !u!114 &-6945963496115081675
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8053848096856963621}
+  m_GameObject: {fileID: 8921437432877574983}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: d5a57f767e5e46a458fc5d3c628d0cbb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  GlobalObjectIdHash: 2409106366
+  GlobalObjectIdHash: 2579471316
   InScenePlacedSourceGlobalObjectIdHash: 0
   AlwaysReplicateAsRoot: 0
   SynchronizeTransform: 1
@@ -156,20 +156,21 @@ MonoBehaviour:
   SpawnWithObservers: 1
   DontDestroyWithOwner: 0
   AutoObjectParentSync: 1
---- !u!114 &-2925161246315457296
+--- !u!114 &-2238994353808586959
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8053848096856963621}
+  m_GameObject: {fileID: 8921437432877574983}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6960c7013aa70cf4e94d1be8f6950f5a, type: 3}
+  m_Script: {fileID: 11500000, guid: d77190dbf4a2255408a880a3bd8dd19e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   timer: 0
-  rb: {fileID: 2075625805928030967}
+  rb: {fileID: 3722686693565726986}
+  kart: {fileID: 0}
   itemCategory: Hazard
   useCount: 1
   isTimed: 0
