+++ GGK/Assets/Prefabs/ItemPrefabs/Hazard/HazardTier3.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 73b47707e9f2e8048828b8709e03b862
+guid: a14434a4591835e49b0ac4efed41f8d4
 PrefabImporter:
   externalObjects: {}
   userData:
