+++ GGK/Assets/Prefabs/ItemPrefabs/Hazard/HazardTier4.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 55aa4467dffa2c24e83ee339dc6ba070
+guid: 73b47707e9f2e8048828b8709e03b862
 PrefabImporter:
   externalObjects: {}
   userData:
