+++ GGK/Assets/Prefabs/ItemPrefabs/Puck/PuckTier1.prefab
@@ -147,7 +147,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d5a57f767e5e46a458fc5d3c628d0cbb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  GlobalObjectIdHash: 3335883840
+  GlobalObjectIdHash: 1291843870
   InScenePlacedSourceGlobalObjectIdHash: 0
   AlwaysReplicateAsRoot: 0
   SynchronizeTransform: 1
@@ -170,6 +170,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   timer: 15
   rb: {fileID: 928149475739924505}
+  kart: {fileID: 0}
   itemCategory: Puck
   useCount: 1
   isTimed: 0
@@ -178,7 +179,6 @@ MonoBehaviour:
   tierTwoItemIcon: {fileID: 0}
   tierThreeItemIcon: {fileID: 0}
   tierFourItemIcon: {fileID: 0}
-  shieldEffect: {fileID: 0}
   currentPos:
     m_InternalValue: {x: 0, y: 0, z: 0}
   maxBounces: 0
