+++ GGK/Assets/Prefabs/MultiplayerNPC.prefab
@@ -20283,7 +20283,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d5a57f767e5e46a458fc5d3c628d0cbb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  GlobalObjectIdHash: 303652607
+  GlobalObjectIdHash: 2445976057
   InScenePlacedSourceGlobalObjectIdHash: 480922830
   AlwaysReplicateAsRoot: 0
   SynchronizeTransform: 1
@@ -35735,6 +35735,7 @@ MonoBehaviour:
   rotationAlignSpeed: 8
   horizontalOffset: 0.21
   doGroundCheck: 1
+  isStunned: 0
   driveSound: {fileID: 0}
   isDriving: 0
   isConfused: 0
@@ -35778,9 +35779,9 @@ MonoBehaviour:
   - {fileID: 7445175988010204192, guid: 246d02636966ef94ca09ee2318afa45c, type: 3}
   HazardArray:
   - {fileID: 7837964673616166884, guid: 6aece729a10356d42ba5111c039d69e6, type: 3}
+  - {fileID: 6181112050242537461, guid: 55aa4467dffa2c24e83ee339dc6ba070, type: 3}
   - {fileID: 7170534580459787604, guid: a14434a4591835e49b0ac4efed41f8d4, type: 3}
   - {fileID: 2161231715911477302, guid: 73b47707e9f2e8048828b8709e03b862, type: 3}
-  - {fileID: 6181112050242537461, guid: 55aa4467dffa2c24e83ee339dc6ba070, type: 3}
   PuckArray:
   - {fileID: 7708316758954177235, guid: cb14a64132daae94a9b5e3beec5ab6c3, type: 3}
   - {fileID: 7708316758954177235, guid: 950b74b59a625df42802ef486dcf0d5d, type: 3}
@@ -35875,6 +35876,7 @@ MonoBehaviour:
   - {fileID: 1909650444857583937}
   - {fileID: 4264326943261783021}
   - {fileID: 50857350959902546}
+  warpBoost: {fileID: 0}
   boostFlames: []
 --- !u!1 &2681580435029215152
 GameObject:
