+++ GGK/Assets/Prefabs/NPCKart.prefab
@@ -20282,7 +20282,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d5a57f767e5e46a458fc5d3c628d0cbb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  GlobalObjectIdHash: 480922830
+  GlobalObjectIdHash: 3814057406
   InScenePlacedSourceGlobalObjectIdHash: 480922830
   AlwaysReplicateAsRoot: 0
   SynchronizeTransform: 1
@@ -35860,6 +35860,7 @@ MonoBehaviour:
   rotationAlignSpeed: 8
   horizontalOffset: 0.21
   doGroundCheck: 1
+  isStunned: 0
   driveSound: {fileID: 0}
   isDriving: 0
   isConfused: 0
@@ -35903,9 +35904,9 @@ MonoBehaviour:
   - {fileID: 7445175988010204192, guid: 246d02636966ef94ca09ee2318afa45c, type: 3}
   HazardArray:
   - {fileID: 7837964673616166884, guid: 6aece729a10356d42ba5111c039d69e6, type: 3}
+  - {fileID: 6181112050242537461, guid: 55aa4467dffa2c24e83ee339dc6ba070, type: 3}
   - {fileID: 7170534580459787604, guid: a14434a4591835e49b0ac4efed41f8d4, type: 3}
   - {fileID: 2161231715911477302, guid: 73b47707e9f2e8048828b8709e03b862, type: 3}
-  - {fileID: 6181112050242537461, guid: 55aa4467dffa2c24e83ee339dc6ba070, type: 3}
   PuckArray:
   - {fileID: 7708316758954177235, guid: cb14a64132daae94a9b5e3beec5ab6c3, type: 3}
   - {fileID: 7708316758954177235, guid: 950b74b59a625df42802ef486dcf0d5d, type: 3}
@@ -36000,6 +36001,7 @@ MonoBehaviour:
   - {fileID: 7430458247879778456}
   - {fileID: 6279542326623655660}
   - {fileID: 4472102584556139945}
+  warpBoost: {fileID: 0}
   boostFlames: []
 --- !u!1 &2681580435029215152
 GameObject:
@@ -122205,6 +122207,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b8e9c182e7220b64abeae9b21225053d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  animator: {fileID: 0}
+  turningValue: 0
   npc: {fileID: 4036233003175977672}
   lerpSpeed: 4
 --- !u!4 &5588149146214448166 stripped
@@ -123150,6 +123154,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b8e9c182e7220b64abeae9b21225053d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  animator: {fileID: 0}
+  turningValue: 0
   npc: {fileID: 4036233003175977672}
   lerpSpeed: 4
 --- !u!4 &5205945830201470416 stripped
@@ -123441,6 +123447,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b8e9c182e7220b64abeae9b21225053d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  animator: {fileID: 0}
+  turningValue: 0
   npc: {fileID: 4036233003175977672}
   lerpSpeed: 4
 --- !u!4 &6638727031424064024 stripped
@@ -123606,6 +123614,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b8e9c182e7220b64abeae9b21225053d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  animator: {fileID: 0}
+  turningValue: 0
   npc: {fileID: 4036233003175977672}
   lerpSpeed: 4
 --- !u!4 &6382769510307900017 stripped
