+++ GGK/Assets/Scenes/StartScene.unity
@@ -1760,7 +1760,7 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 628999392}
   m_HandleRect: {fileID: 628999391}
   m_Direction: 2
-  m_Value: 0
+  m_Value: 1
   m_Size: 1
   m_NumberOfSteps: 0
   m_OnValueChanged:
@@ -3778,8 +3778,7 @@ RectTransform:
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
-  m_Children:
-  - {fileID: 1254118460}
+  m_Children: []
   m_Father: {fileID: 444550953}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
@@ -4352,11 +4351,11 @@ RectTransform:
   m_LocalScale: {x: 1.0561956, y: 1.0561956, z: 1.0561956}
   m_ConstrainProportionsScale: 0
   m_Children: []
-  m_Father: {fileID: 1254118460}
+  m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -0.00090026855, y: 135}
+  m_AnchoredPosition: {x: -0.00090026855, y: 1058.4666}
   m_SizeDelta: {x: 650, y: 0}
   m_Pivot: {x: 0.5, y: 1}
 --- !u!114 &1567664993
@@ -5587,3 +5586,4 @@ SceneRoots:
   - {fileID: 376136322}
   - {fileID: 1011990453}
   - {fileID: 1499987440}
+  - {fileID: 1567664992}
