+++ GGK/Assets/Scripts/ItemScripts/Hazards/HazardTier2.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d77190dbf4a2255408a880a3bd8dd19e
+guid: 8044fcaf0c56b1f448367db2d69bfa64
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
