+++ GGK/Assets/Scripts/ItemScripts/Hazards/HazardTier3.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6960c7013aa70cf4e94d1be8f6950f5a
+guid: d77190dbf4a2255408a880a3bd8dd19e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
