+++ GGK/Assets/Scripts/ItemScripts/Hazards/HazardTier4.cs
@@ -1,54 +1,60 @@
 using System.Collections;
 using System.Collections.Generic;
-using Unity.VisualScripting;
+using Unity.Netcode;
 using UnityEngine;
-using UnityEngine.UIElements;
 
 /// <summary>
-/// Fake Item Box
+/// Blizzard
 /// </summary>
 public class HazardTier4 : BaseItem
 {
-    // Start is called before the first frame update
-    void Start()
+    private void Start()
     {
-        timer = 10.0f;
-
-        // sends the hazard slightly up and behind the player before landing on the ground
-        transform.position = transform.position
-                             - transform.forward * 5f   // behind the kart
-                             + transform.up * 1.5f;       // slightly above ground
+        Vector3 behindPos = transform.position - transform.forward * 8;
+        behindPos.y += 3.0f;
+        transform.position = behindPos;
     }
 
-    // Update is called once per frame
     private new void Update()
     {
         RotateBox();
-        DecreaseTimer();
     }
 
-    public void RotateBox()
+    private void RotateBox()
     {
         transform.rotation *= new Quaternion(0.0f, 2.0f * Time.deltaTime, 0.0f, 1.0f);
     }
 
     private void OnTriggerEnter(Collider collision)
     {
+        // stop the trap from falling when they reach the ground/road
+        if (collision.gameObject.CompareTag("Ground") || collision.gameObject.CompareTag("Road"))
+        {
+            rb.constraints = RigidbodyConstraints.FreezePositionY;
+        }
         if (collision.gameObject.CompareTag("Kart")) // checks if kart gameobject player or npc
         {
             Rigidbody kartRigidbody;
             if (collision.gameObject.TryGetComponent<Rigidbody>(out kartRigidbody)) // checks if they have rb while also assigning if they do
             {
                 kartRigidbody.velocity *= 0.125f; //this slows a kart down to an eighth of its speed
-
-                if (!MultiplayerManager.Instance.IsMultiplayer)
-                {
-                    Destroy(this.gameObject);
-                }
-                else
-                {
-                    DestroyItemRpc(this);
-                }
+            }
+            NEWDriver playerKart = collision.gameObject.gameObject.GetComponentInChildren<NEWDriver>();
+            if (playerKart)
+            {
+                playerKart.confusedTimer = 10;
+                playerKart.isConfused = true;
+                playerKart.movementDirection *= -1;
+            }
+            
+            // destroy puck if single player, if multiplayer call rpc in base item to destroy and despawn
+            if (!MultiplayerManager.Instance.IsMultiplayer)
+            {
+                Destroy(this.gameObject);
+            }
+            else
+            {
+                DestroyItemRpc(this);
             }
         }
     }
