+++ GGK/Assets/Scripts/ItemScripts/Hazards/HazardTier4.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8044fcaf0c56b1f448367db2d69bfa64
+guid: 6960c7013aa70cf4e94d1be8f6950f5a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
